+++ src/test/integration/DepartmentTest.java
@@ -41,33 +41,4 @@ public class DepartmentTest {
         adminController.declineDepartmentChange(user);
         Assert.assertEquals("Department is not changed", Department.DEVELOPER, user.getDepartment());
     }
-
-
-    @Test
-    public void testRequestSalaryAndApply() {
-        Assert.assertEquals("User salary is incorrect", 10, user.getSalary());
-
-        userController.submitSalaryChange(user, 20);
-        Assert.assertEquals("Requested salary not changed", 20, user.getRequestedSalary());
-        AccountList.accountArrayList.add(user);
-
-        Assert.assertTrue("Incorrect any Requests returned with new salary requested", adminController.displayAllRequests());
-
-        adminController.acceptSalaryChange(user);
-        Assert.assertEquals("Salary is not changed", 20, user.getSalary());
-    }
-
-    @Test
-    public void testRequestSalaryAndDecline() {
-        Assert.assertEquals("User salary is incorrect", 10, user.getSalary());
-
-        userController.submitSalaryChange(user, 20);
-        Assert.assertEquals("Requested salary not changed", 20, user.getRequestedSalary());
-        AccountList.accountArrayList.add(user);
-
-        Assert.assertTrue("Incorrect any Requests returned with new salary requested", adminController.displayAllRequests());
-
-        adminController.declineSalaryChange(user);
-        Assert.assertEquals("Salary is not changed", 10, user.getSalary());
-    }
 }
